+++ test/pg-structure.js
@@ -18,7 +18,8 @@ before(function (done) {
 });
 
 after(function (done) {
-    testDB.dropDB(done);
+    //testDB.dropDB(done);
+    done();
 });
 
 describe('DB Object', function () {
@@ -47,5 +48,26 @@ describe('User Defined Type', function () {
     });
     it('should have Sequelize Type', function () {
         assert.equal(db.schema('public').table('type_table').column('company').sequelizeType(), 'DataTypes.STRING');
+    });
+});
+
+describe('Column Default', function () {
+    it('contact should have some default values.', function () {
+        assert.equal(db.schema('public').table('contact').column('name').default(), '\'oz\'');
+        assert.equal(db.schema('public').table('contact').column('surname').default(), '\'O\'\'Reilly\'');
+        assert.equal(db.schema('public').table('contact').column('birth_date').default(), '\'2010-01-01\'');
+        assert.equal(db.schema('public').table('contact').column('is_active').default(), 'true');
+        assert.equal(db.schema('public').table('contact').column('email').default(), '\'x@x.com\'');
+    });
+    it('contact should have some default values with typecast.', function () {
+        assert.equal(db.schema('public').table('contact').column('name').defaultWithTypeCast(), '\'oz\'::character varying');
+        assert.equal(db.schema('public').table('contact').column('surname').defaultWithTypeCast(), '\'O\'\'Reilly\'::character varying');
+        assert.equal(db.schema('public').table('contact').column('birth_date').defaultWithTypeCast(), '\'2010-01-01\'::date');
+        assert.equal(db.schema('public').table('contact').column('is_active').defaultWithTypeCast(), 'true');
+        assert.equal(db.schema('public').table('contact').column('email').defaultWithTypeCast(), '\'x@x.com\'::character varying');
+    });
+
+    it('cart_line_item should have some default values.', function () {
+        assert.equal(db.schema('public').table('cart_line_item').column('quantity').default(), 1);
     });
 });