+++ History.md
@@ -10,6 +10,12 @@ Note
 Version history for minimal documentation updates are not listed here to prevent cluttering.
 Important documentation changes are included anyway.
 
+1.5.0 / 2014-12-04
+==================
+* Fix: column.default() method returns default value with type cast. From now on yype cast part is stripped.
+* Added column.defaultWithTypeCast() method for getting default values with type cast part.
+* Tests added for default values.
+
 1.4.0 / 2014-11-28
 ==================
 * Added support for user-defined types.